# src/config/swagger.yml
swagger: "2.0"
info:
  description: "Flood api to register and list floods"
  version: "1.0.0"
  title: "wave"
basePath: "/api/v1"
consumes:
  - "application/json"
  - "application/octet-stream"
produces:
  - "application/json"
tags:
  - name: "Users"
    description: "User Endpoints"
  - name: "Flood"
    description: "Flood endpoints endpoints"
paths:
  /create-user:
    post:
      summary: User Registration
      tags: 
        - "Users"
      x-swagger-router-controller: app
      operationId: createUserHandler
      parameters:
        - name: body
          in: body
          description: "Object containing user details"
          schema: 
            type: object
            properties:
              first_name:
                type: string
              last_name:
                type: string
              username:
                type: string
              password:
                type: string
              email:
                type: string
              user_location:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
          required: true
      responses:
        '200':
          description: User registered successfully
          schema:
            type: object
            properties:
              id:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              username:
                type: string
        '500':
          description: Internal server error

  /login:
    post:
      summary: User Login
      tags: 
        - "Users"
      x-swagger-router-controller: app
      operationId: loginHandler
      parameters:
        - name: body
          in: body
          description: "Object containing user details"
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              
      responses:
        '200':
          description: Login successful
          schema:
            type: object
            properties:
              token:
                type: string
        '500':
          description: Internal server error
  /floods:
    get:
      summary: List of Floods
      tags: 
        - "Flood"
      x-swagger-router-controller: app
      operationId: listFloodHandler
      parameters:
        - in: query
          name: latitude
          type: number
          description: Latitude for geolocation filtering
        - in: query
          name: longitude
          type: number
          description: Longitude for geolocation filtering
      responses:
        '200':
          description: List of floods retrieved successfully
          schema:
            type: array
            items:
              type: object
              properties:
                flood_id:
                  type: string
                user_id:
                  type: string
                flood_center_location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                user_location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                start_date:
                  type: string
                end_date:
                  type: string
                water_level:
                  type: number
                description:
                  type: array
                  items:
                    type: string
                flood_radius:
                  type: number
                users:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                status:
                  type: string
        '500':
          description: Internal server error
  /create-flood:
    post:
      summary: Flood Registration
      tags: 
        - "Flood"
      x-swagger-router-controller: app
      operationId: createFloodHandler
      parameters:
        - name: body
          in: body
          description: "Object containing user details"
          required: true
          schema:
              type: object
              properties:
                flood_center_location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                user_location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                date_time:
                  type: string
                water_level:
                  type: number
                description:
                  type: string
                attachments:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Flood registered successfully
          schema:
            type: object
            properties:
              flood_id:
                type: string
                format: uuid
        '500':
          description: Internal server error

  /detailed-flood:
    get:
      summary: Flood Details
      tags: 
        - "Flood"
      x-swagger-router-controller: app
      operationId: detailedFloodHandler
      parameters:
        - in: query
          name: flood_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Flood details retrieved successfully
          schema:
            type: object
            properties:
              flood_id:
                type: string
                format: uuid
              user_id:
                type: string
                format: uuid
              flood_center_location:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
              user_location:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
              start_date:
                  type: string
              end_date:
                  type: string
              water_level:
                type: number
              description:
                type: string
              flood_radius:
                type: number
              users:
                type: array
                items:
                  type: object
                  properties:
                    user_id:
                      type: string
                    first_name:
                      type: string
              status:
                type: string
        '500':
          description: Internal server error
  /update-flood:
    put:
      summary: Flood Update
      tags: 
        - "Flood"
      x-swagger-router-controller: app
      operationId: updateFloodHandler
      parameters:
        - in: query
          name: flood_id
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          description: "Object containing user details"
          required: true
          schema:
            type: object
            properties:
              user_location:
                type: object
                properties:
                  latitude:
                    type: number
                  longitude:
                    type: number
              date_time:
                type: string
              water_level:
                type: number
              description:
                type: string
              status:
                type: string
      responses:
        '200':
          description: Flood updated successfully
          schema:
            type: object
            properties:
              flood_id:
                type: string
                format: uuid
        '500':
          description: Internal server error
  /delete-flood:
    delete:
      summary: Flood Deletion
      tags: 
        - "Flood"
      x-swagger-router-controller: app
      operationId: deleteFloodHandler
      parameters:
        - in: query
          name: flood_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Flood deleted successfully
        '500':
          description: Internal server error